refactor: split frontend validation helpers
